Fix integer types in #1114

+++ include/CoolPropLib.h
@@ -366,7 +366,7 @@
     * @param buffer_length The length of the buffer for the error code
     * @return
     */
-    EXPORT_CODE void CONVENTION AbstractState_set_binary_interaction_double(const long handle, const size_t i, const size_t j, const char* parameter, const double value, long *errcode, char *message_buffer, const long buffer_length);
+    EXPORT_CODE void CONVENTION AbstractState_set_binary_interaction_double(const long handle, const long i, const long j, const char* parameter, const double value, long *errcode, char *message_buffer, const long buffer_length);
 
     // *************************************************************************************
     // *************************************************************************************
