+++ src/CoolPropLib.cpp
@@ -648,7 +648,7 @@ EXPORT_CODE void CONVENTION AbstractState_update_and_5_out(const long handle, co
     }
 }
 
-EXPORT_CODE void CONVENTION AbstractState_set_binary_interaction_double(const long handle, const size_t i, const size_t j, const char* parameter, const double value, long *errcode, char *message_buffer, const long buffer_length)
+EXPORT_CODE void CONVENTION AbstractState_set_binary_interaction_double(const long handle, const long i, const long j, const char* parameter, const double value, long *errcode, char *message_buffer, const long buffer_length)
 {
     *errcode = 0;
     try {
